requirements.txt
aiodns==3.0.0
aiogram==2.24
aiohttp==3.8.3
aiosignal==1.3.1
amqp==2.6.1
anyio==3.6.2
argon2-cffi==19.1.0
asgiref==3.6.0
astroid==2.13.3
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.11.2
billiard==3.6.4.0
binance==0.3
binance-connector==2.0.0
binance-connector-python==0.6.0
blinker==1.4
Brotli==1.0.9
bs4==0.0.1
celery==4.4.2
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.1
dateparser==1.1.7
dbus-python==1.2.18
deepdiff==6.2.3
defusedxml==0.7.1
distro==1.7.0
Django==3.0.4
django-allauth==0.53.1
django-coverage-plugin==1.6.0
django-environ==0.4.5
django-extensions==2.1.7
django-filter==2.2.0
django-model-utils==3.1.2
djangorestframework==3.11.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.3
environs==9.5.0
Faker==12.0.1
fastapi==0.89.1
fasteners==0.14.1
flake8==3.7.5
frozenlist==1.3.3
future==0.18.2
h11==0.14.0
httpagentparser==1.9.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.6.4
incremental==22.10.0
iniconfig==1.1.1
isort==4.3.21
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==3.2.0
keyring==23.5.0
kombu==4.6.11
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lockfile==0.12.2
lxml==4.9.2
macaroonbakery==1.3.1
magic-filter==1.0.9
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.19.0
mccabe==0.6.1
mixer==7.2.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
ordered-set==4.1.0
orjson==3.8.5
packaging==21.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.1
protobuf==3.12.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
py==1.10.0
py-currency-converter==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycares==4.3.0
pycodestyle==2.5.0
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.17
pycups==2.0.1
pydantic==1.10.4
pyflakes==2.1.1
Pygments==2.4.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyotp==2.8.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==5.2.1
pytest-cov==2.8.1
pytest-django==3.5.1
pytest-mock==1.11.1
pytest-sugar==0.9.2
python-binance==1.0.16
python-dateutil==2.8.0
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-slugify==3.0.4
python3-openid==3.2.0
pytz==2019.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
redis==4.4.2
regex==2022.10.31
reportlab==3.6.8
requests==2.26.0
requests-oauthlib==1.3.1
SecretStorage==3.3.1
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
starlette==0.22.0
stripe==5.2.0
systemd-python==234
termcolor==2.2.0
text-unidecode==1.3
texttable==1.6.4
toml==0.10.2
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.24.3
usb-creator==0.3.7
uvicorn==0.20.0
vine==1.3.0
wadllib==1.3.6
wcwidth==0.2.6
websocket-client==1.2.3
websockets==10.4
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.2
zipp==1.0.0
zope.interface==5.5.2
